+++ example/windows/flutter/generated_plugin_registrant.cc
@@ -6,12 +6,12 @@
 
 #include "generated_plugin_registrant.h"
 
-#include <notification_manager/notification_manager_plugin_c_api.h>
+#include <flutter_system_notifications/flutter_system_notifications_plugin_c_api.h>
 #include <permission_handler_windows/permission_handler_windows_plugin.h>
 
 void RegisterPlugins(flutter::PluginRegistry* registry) {
-  NotificationManagerPluginCApiRegisterWithRegistrar(
-      registry->GetRegistrarForPlugin("NotificationManagerPluginCApi"));
+  FlutterSystemNotificationsPluginCApiRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("FlutterSystemNotificationsPluginCApi"));
   PermissionHandlerWindowsPluginRegisterWithRegistrar(
       registry->GetRegistrarForPlugin("PermissionHandlerWindowsPlugin"));
 }